+++ lib/twilio/connection.ts
@@ -7,11 +7,11 @@ import { EventEmitter } from 'events';
 import Device from './device';
 import DialtonePlayer from './dialtonePlayer';
 import { GeneralErrors, InvalidArgumentError, MediaErrors, TwilioError } from './errors';
+import Logger from './logger';
 import { RTCIceCandidate, RTCLocalIceCandidate } from './rtc/candidate';
 import RTCSample from './rtc/sample';
 import RTCWarning from './rtc/warning';
 import StatsMonitor from './statsMonitor';
-import Log, { LogLevel } from './tslog';
 import { isChrome } from './util';
 
 const Backoff = require('backoff');
@@ -168,9 +168,9 @@ class Connection extends EventEmitter {
   private _latestOutputVolume: number = 0;
 
   /**
-   * An instance of Log to use.
+   * An instance of Logger to use.
    */
-  private readonly _log: Log = new Log(LogLevel.Off);
+  private _logger: Logger = Logger.getInstance();
 
   /**
    * An instance of Backoff for media reconnection
@@ -222,7 +222,6 @@ class Connection extends EventEmitter {
    * Options passed to this {@link Connection}.
    */
   private options: Connection.Options = {
-    debug: false,
     enableRingingState: false,
     mediaStreamFactory: PeerConnection,
     offerSdp: null,
@@ -262,10 +261,6 @@ class Connection extends EventEmitter {
 
     this._direction = this.parameters.CallSid ? Connection.CallDirection.Incoming : Connection.CallDirection.Outgoing;
 
-    this._log.setLogLevel(this.options.debug ? LogLevel.Debug
-      : this.options.warnings ? LogLevel.Warn
-      : LogLevel.Off);
-
     this._mediaReconnectBackoff = Backoff.exponential(BACKOFF_CONFIG);
     this._mediaReconnectBackoff.on('ready', () => this.mediaStream.iceRestart());
 
@@ -295,13 +290,11 @@ class Connection extends EventEmitter {
     this.mediaStream = new (this.options.MediaStream || this.options.mediaStreamFactory)
       (config.audioHelper, config.pstream, config.getUserMedia, {
         codecPreferences: this.options.codecPreferences,
-        debug: this.options.debug,
         dscp: this.options.dscp,
         enableIceRestart: this.options.enableIceRestart,
         forceAggressiveIceNomination: this.options.forceAggressiveIceNomination,
         isUnifiedPlan: this._isUnifiedPlanDefault,
         maxAverageBitrate: this.options.maxAverageBitrate,
-        warnings: this.options.warnings,
       });
 
     this.on('volume', (inputVolume: number, outputVolume: number): void => {
@@ -363,7 +356,7 @@ class Connection extends EventEmitter {
     };
 
     this.mediaStream.ondisconnected = (msg: string): void => {
-      this._log.info(msg);
+      this._logger.info(msg);
       this._publisher.warn('network-quality-warning-raised', 'ice-connectivity-lost', {
         message: msg,
       }, this);
@@ -384,7 +377,7 @@ class Connection extends EventEmitter {
     };
 
     this.mediaStream.onreconnected = (msg: string): void => {
-      this._log.info(msg);
+      this._logger.info(msg);
       this._publisher.info('network-quality-warning-cleared', 'ice-connectivity-lost', {
         message: msg,
       }, this);
@@ -404,7 +397,7 @@ class Connection extends EventEmitter {
         twilioError: e.info.twilioError,
       };
 
-      this._log.error('Received an error from MediaStream:', e);
+      this._logger.error('Received an error from MediaStream:', e);
       this.emit('error', error);
     };
 
@@ -448,7 +441,7 @@ class Connection extends EventEmitter {
     this.pstream.on('ringing', this._onRinging);
 
     this.pstream.on('transportClose', () => {
-      this._log.error('Received transportClose from pstream');
+      this._logger.error('Received transportClose from pstream');
       this.emit('transportClose');
     });
 
@@ -473,7 +466,7 @@ class Connection extends EventEmitter {
    * @private
    */
   _getRealCallSid(): string | null {
-    this._log.warn('_getRealCallSid is deprecated and will be removed in 2.0.');
+    this._logger.warn('_getRealCallSid is deprecated and will be removed in 2.0.');
     return /^TJ/.test(this.parameters.CallSid) ? null : this.parameters.CallSid;
   }
 
@@ -483,7 +476,7 @@ class Connection extends EventEmitter {
    * @private
    */
   _getTempCallSid(): string | undefined {
-    this._log.warn('_getTempCallSid is deprecated and will be removed in 2.0. \
+    this._logger.warn('_getTempCallSid is deprecated and will be removed in 2.0. \
                     Please use outboundConnectionId instead.');
     return this.outboundConnectionId;
   }
@@ -638,7 +631,7 @@ class Connection extends EventEmitter {
    */
   cancel(handler: () => void): void;
   cancel(handler?: () => void): void {
-    this._log.warn('.cancel() is deprecated. Please use .ignore() instead.');
+    this._logger.warn('.cancel() is deprecated. Please use .ignore() instead.');
 
     if (handler) {
       this.ignore(handler);
@@ -847,7 +840,7 @@ class Connection extends EventEmitter {
 
     if (dtmfSender) {
       if (!('canInsertDTMF' in dtmfSender) || dtmfSender.canInsertDTMF) {
-        this._log.info('Sending digits using RTCDTMFSender');
+        this._logger.info('Sending digits using RTCDTMFSender');
         // NOTE(mroberts): We can't just map 'w' to ',' since
         // RTCDTMFSender's pause duration is 2 s and Twilio's is more
         // like 500 ms. Instead, we will fudge it with setTimeout.
@@ -855,11 +848,11 @@ class Connection extends EventEmitter {
         return;
       }
 
-      this._log.info('RTCDTMFSender cannot insert DTMF');
+      this._logger.info('RTCDTMFSender cannot insert DTMF');
     }
 
     // send pstream message to send DTMF
-    this._log.info('Sending digits over PStream');
+    this._logger.info('Sending digits over PStream');
 
     if (this.pstream !== null && this.pstream.status !== 'disconnected') {
       this.pstream.publish('dtmf', {
@@ -893,7 +886,7 @@ class Connection extends EventEmitter {
    * @deprecated - Unmute the {@link Connection}.
    */
   unmute(): void {
-    this._log.warn('.unmute() is deprecated. Please use .mute(false) to unmute a call instead.');
+    this._logger.warn('.unmute() is deprecated. Please use .mute(false) to unmute a call instead.');
     this.mute(false);
   }
 
@@ -903,7 +896,7 @@ class Connection extends EventEmitter {
    */
   volume(handler: (inputVolume: number, outputVolume: number) => void): void {
     if (!window || (!(window as any).AudioContext && !(window as any).webkitAudioContext)) {
-      this._log.warn('This browser does not support Connection.volume');
+      this._logger.warn('This browser does not support Connection.volume');
     }
 
     this._addHandler('volume', handler);
@@ -916,7 +909,7 @@ class Connection extends EventEmitter {
    */
   private _addHandler(eventName: string, handler: (...args: any[]) => any): this {
     if (!hasBeenWarnedHandlers) {
-      this._log.warn(`Connection callback handlers (accept, cancel, disconnect, error, ignore, mute, reject,
+      this._logger.warn(`Connection callback handlers (accept, cancel, disconnect, error, ignore, mute, reject,
         volume) have been deprecated and will be removed in the next breaking release. Instead, the EventEmitter \
         interface can be used to set event listeners. Example: connection.on('${eventName}', handler)`);
       hasBeenWarnedHandlers = true;
@@ -1021,7 +1014,7 @@ class Connection extends EventEmitter {
       return;
     }
 
-    this._log.info('Disconnecting...');
+    this._logger.info('Disconnecting...');
 
     // send pstream hangup message
     if (this.pstream !== null && this.pstream.status !== 'disconnected' && this.sendHangup) {
@@ -1146,7 +1139,7 @@ class Connection extends EventEmitter {
       return;
     }
 
-    this._log.info('Received HANGUP from gateway');
+    this._logger.info('Received HANGUP from gateway');
     if (payload.error) {
       const error = {
         code: payload.error.code || 31000,
@@ -1154,7 +1147,7 @@ class Connection extends EventEmitter {
         message: payload.error.message || 'Error sent from gateway in HANGUP',
         twilioError: new GeneralErrors.ConnectionError(),
       };
-      this._log.error('Received an error from the gateway:', error);
+      this._logger.error('Received an error from the gateway:', error);
       this.emit('error', error);
     }
     this.sendHangup = false;
@@ -1200,7 +1193,7 @@ class Connection extends EventEmitter {
 
         // We already exceeded max retry time.
         if (Date.now() - this._mediaReconnectStartTime > BACKOFF_CONFIG.maxDelay) {
-          this._log.info('Exceeded max ICE retries');
+          this._logger.info('Exceeded max ICE retries');
           return this.mediaStream.onerror(MEDIA_DISCONNECT_ERROR);
         }
 
@@ -1226,7 +1219,7 @@ class Connection extends EventEmitter {
         message: 'Media connection failed.',
         twilioError: new MediaErrors.ConnectionError(),
       };
-      this._log.warn('ICE Connection disconnected.');
+      this._logger.warn('ICE Connection disconnected.');
       this._publisher.warn('connection', 'error', mediaReconnectionError, this);
       this._publisher.info('connection', 'reconnecting', null, this);
 
@@ -1248,7 +1241,7 @@ class Connection extends EventEmitter {
     if (this._status !== Connection.State.Reconnecting) {
       return;
     }
-    this._log.info('ICE Connection reestablished.');
+    this._logger.info('ICE Connection reestablished.');
     this._publisher.info('connection', 'reconnected', null, this);
 
     this._status = Connection.State.Open;
@@ -1320,7 +1313,7 @@ class Connection extends EventEmitter {
     this._publisher.postMetrics(
       'quality-metrics-samples', 'metrics-sample', this._metricsSamples.splice(0), this._createMetricPayload(), this,
     ).catch((e: any) => {
-      this._log.warn('Unable to post metrics to Insights. Received error:', e);
+      this._logger.warn('Unable to post metrics to Insights. Received error:', e);
     });
   }
 
@@ -1593,11 +1586,6 @@ namespace Connection {
     codecPreferences?: Codec[];
 
     /**
-     * Whether to enable debug level logging.
-     */
-    debug?: boolean;
-
-    /**
      * A DialTone player, to play mock DTMF sounds.
      */
     dialtonePlayer?: DialtonePlayer;
@@ -1698,11 +1686,6 @@ namespace Connection {
      * TwiML params for the call. May be set for either outgoing or incoming calls.
      */
     twimlParams?: Record<string, any>;
-
-    /**
-     * Whether to enable warn level logging.
-     */
-    warnings?: boolean;
   }
 
   /**